# specs/openapi/admin-v0.4.2.json
{
  "components": {
    "schemas": {
...
      "EffectiveNamespaceGovernanceOut": {
        "properties": {
          "disclosures": {
            "items": {
              "type": "string"
            },
            "title": "Disclosures",
            "type": "array"
          },
          "enforcement_basis": {
            "const": "server_configuration",
            "default": "server_configuration",
            "title": "Enforcement Basis",
            "type": "string"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "policy": {
            "$ref": "#/components/schemas/NamespaceGovernancePolicyOut"
          },
          "usage": {
            "$ref": "#/components/schemas/NamespaceDailyUsageOut"
          }
        },
        "required": [
          "generated_at",
          "policy",
          "usage",
          "disclosures"
        ],
        "title": "EffectiveNamespaceGovernanceOut",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NamespaceGovernancePolicyOut": {
        "properties": {
          "allowed_processing_regions": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Allowed Processing Regions"
          },
          "allowed_recorder_capture_modes": {
            "anyOf": [
              {
                "items": {
                  "enum": [
                    "metadata_only",
                    "hash_only",
                    "full"
                  ],
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Allowed Recorder Capture Modes"
          },
          "configured": {
            "title": "Configured",
            "type": "boolean"
          },
          "decision_records_daily_limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision Records Daily Limit"
          },
          "deployment_region": {
            "title": "Deployment Region",
            "type": "string"
          },
          "effective_recorder_capture_modes": {
            "items": {
              "enum": [
                "metadata_only",
                "hash_only",
                "full"
              ],
              "type": "string"
            },
            "title": "Effective Recorder Capture Modes",
            "type": "array"
          },
          "estimated_ingest_bytes_daily_limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Ingest Bytes Daily Limit"
          },
          "governance_created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Governance Created At"
          },
          "governance_created_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Governance Created By"
          },
          "governance_updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Governance Updated At"
          },
          "governance_updated_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Governance Updated By"
          },
          "memory_writes_daily_limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Memory Writes Daily Limit"
          },
          "namespace": {
            "title": "Namespace",
            "type": "string"
          },
          "policy_version": {
            "title": "Policy Version",
            "type": "integer"
          },
          "processing_region_allowed": {
            "title": "Processing Region Allowed",
            "type": "boolean"
          },
          "recalls_daily_limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recalls Daily Limit"
          },
          "recorder_events_daily_limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recorder Events Daily Limit"
          },
          "status": {
            "enum": [
              "unconfigured",
              "active",
              "disabled"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "namespace",
          "configured",
          "status",
          "policy_version",
          "deployment_region",
          "processing_region_allowed",
          "allowed_processing_regions",
          "allowed_recorder_capture_modes",
          "effective_recorder_capture_modes",
          "recorder_events_daily_limit",
          "decision_records_daily_limit",
          "memory_writes_daily_limit",
          "recalls_daily_limit",
          "estimated_ingest_bytes_daily_limit",
          "governance_created_at",
          "governance_created_by",
          "governance_updated_at",
          "governance_updated_by"
        ],
        "title": "NamespaceGovernancePolicyOut",
        "type": "object"
      },
      "NamespaceGovernancePolicyUpdate": {
        "additionalProperties": false,
        "description": "Complete replacement of the enforceable portion of a namespace policy.",
        "properties": {
          "allowed_processing_regions": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "maxItems": 100,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Allowed Processing Regions"
          },
          "allowed_recorder_capture_modes": {
            "anyOf": [
              {
                "items": {
                  "enum": [
                    "metadata_only",
                    "hash_only",
                    "full"
                  ],
                  "type": "string"
                },
                "maxItems": 3,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Allowed Recorder Capture Modes"
          },
          "decision_records_daily_limit": {
            "anyOf": [
              {
                "maximum": 9.223372036854776e+18,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision Records Daily Limit"
          },
          "estimated_ingest_bytes_daily_limit": {
            "anyOf": [
              {
                "maximum": 9.223372036854776e+18,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Ingest Bytes Daily Limit"
          },
          "expected_version": {
            "description": "Current policy_version, or 0 to assert first configuration",
            "minimum": 0.0,
            "title": "Expected Version",
            "type": "integer"
          },
          "memory_writes_daily_limit": {
            "anyOf": [
              {
                "maximum": 9.223372036854776e+18,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Memory Writes Daily Limit"
          },
          "recalls_daily_limit": {
            "anyOf": [
              {
                "maximum": 9.223372036854776e+18,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recalls Daily Limit"
          },
          "recorder_events_daily_limit": {
            "anyOf": [
              {
                "maximum": 9.223372036854776e+18,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recorder Events Daily Limit"
          }
        },
        "required": [
          "expected_version"
        ],
        "title": "NamespaceGovernancePolicyUpdate",
        "type": "object"
      },
      "NamespaceGovernanceStatusOut": {
        "properties": {
          "effective": {
            "$ref": "#/components/schemas/EffectiveNamespaceGovernanceOut"
          },
          "latest_snapshot_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Snapshot Hash"
          },
          "revision_count": {
            "title": "Revision Count",
            "type": "integer"
          }
        },
        "required": [
          "effective",
          "revision_count",
          "latest_snapshot_hash"
        ],
        "title": "NamespaceGovernanceStatusOut",
        "type": "object"
      },
      "NamespaceGovernanceStatusUpdate": {
        "additionalProperties": false,
        "properties": {
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          },
          "status": {
            "enum": [
              "active",
              "disabled"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "expected_version",
          "status"
        ],
        "title": "NamespaceGovernanceStatusUpdate",
        "type": "object"
      },
      "NamespacePolicyRevisionOut": {
        "properties": {
          "action": {
            "enum": [
              "created",
              "updated",
              "enabled",
              "disabled",
              "cleared"
            ],
            "title": "Action",
            "type": "string"
          },
          "actor_id": {
            "title": "Actor Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "namespace": {
            "title": "Namespace",
            "type": "string"
          },
          "policy_snapshot": {
            "additionalProperties": true,
            "title": "Policy Snapshot",
            "type": "object"
          },
          "policy_version": {
            "title": "Policy Version",
            "type": "integer"
          },
          "snapshot_hash": {
            "title": "Snapshot Hash",
            "type": "string"
          }
        },
        "required": [
          "namespace",
          "policy_version",
          "action",
          "actor_id",
          "policy_snapshot",
          "snapshot_hash",
          "created_at"
        ],
        "title": "NamespacePolicyRevisionOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "in": "header",
        "name": "X-Admin-Secret",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Provider-neutral decision receipts, bitemporal reconstruction, change-impact analysis, and auditable agent memory",
    "title": "Lians Decision Evidence API",
    "version": "0.4.2",
    "x-lians-api-surface": "admin"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/admin/governance/policies": {
      "get": {
        "operationId": "admin_list_policies_v1_admin_governance_policies_get",
        "parameters": [
          {
            "in": "query",
            "name": "include_unconfigured",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Unconfigured",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/NamespaceGovernancePolicyOut"
                  },
                  "title": "Response Admin List Policies V1 Admin Governance Policies Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "List namespace governance policies",
        "tags": [
          "admin",
          "governance"
        ]
      }
    },
    "/v1/admin/governance/policies/{namespace}": {
      "delete": {
        "operationId": "admin_clear_policy_v1_admin_governance_policies__namespace__delete",
        "parameters": [
          {
            "in": "path",
            "name": "namespace",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]{0,254}$",
              "title": "Namespace",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "expected_version",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Expected Version",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-Admin-Actor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Actor"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Clear governance fields while preserving retention and billing policy",
        "tags": [
          "admin",
          "governance"
        ]
      },
      "get": {
        "operationId": "admin_get_policy_v1_admin_governance_policies__namespace__get",
        "parameters": [
          {
            "in": "path",
            "name": "namespace",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]{0,254}$",
              "title": "Namespace",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NamespaceGovernancePolicyOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Get one configured namespace governance policy",
        "tags": [
          "admin",
          "governance"
        ]
      },
      "put": {
        "operationId": "admin_put_policy_v1_admin_governance_policies__namespace__put",
        "parameters": [
          {
            "in": "path",
            "name": "namespace",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]{0,254}$",
              "title": "Namespace",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Admin-Actor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Actor"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NamespaceGovernancePolicyUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NamespaceGovernancePolicyOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Create or completely replace a namespace governance policy",
        "tags": [
          "admin",
          "governance"
        ]
      }
    },
    "/v1/admin/governance/policies/{namespace}/revisions": {
      "get": {
        "operationId": "admin_list_policy_revisions_v1_admin_governance_policies__namespace__revisions_get",
        "parameters": [
          {
            "in": "path",
            "name": "namespace",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]{0,254}$",
              "title": "Namespace",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/NamespacePolicyRevisionOut"
                  },
                  "title": "Response Admin List Policy Revisions V1 Admin Governance Policies  Namespace  Revisions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "List immutable governance policy revisions",
        "tags": [
          "admin",
          "governance"
        ]
      }
    },
    "/v1/admin/governance/policies/{namespace}/status": {
      "put": {
        "operationId": "admin_set_policy_status_v1_admin_governance_policies__namespace__status_put",
        "parameters": [
          {
            "in": "path",
            "name": "namespace",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]{0,254}$",
              "title": "Namespace",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Admin-Actor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Actor"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NamespaceGovernanceStatusUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NamespaceGovernancePolicyOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Enable or disable a configured namespace governance policy",
        "tags": [
          "admin",
          "governance"
        ]
      }
    },
    "/v1/admin/governance/status/{namespace}": {
      "get": {
        "operationId": "admin_governance_status_v1_admin_governance_status__namespace__get",
        "parameters": [
          {
            "in": "path",
            "name": "namespace",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]{0,254}$",
              "title": "Namespace",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NamespaceGovernanceStatusOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Inspect effective enforcement, usage, and immutable revision state",
        "tags": [
          "admin",
          "governance"
        ]
      }
    },
...
  }
}